+++ pycricinfo/api/endpoints/seasons.py
@@ -2,7 +2,8 @@ from typing import Literal
 
 from fastapi import APIRouter, Path, Query, status
 
-from pycricinfo.search.seasons import extract_match_ids_from_series, get_match_types_in_season
+from pycricinfo.search.seasons import get_match_types_in_season
+from pycricinfo.search.series import extract_match_ids_from_series
 from pycricinfo.source_models.pages.series import MatchResult, MatchType
 
 router = APIRouter(prefix="/matches", tags=["matches"])