+++ pycricinfo/api/server.py
@@ -1,9 +1,11 @@
 import uvicorn
-from fastapi import FastAPI
+from fastapi import FastAPI, Request
+from fastapi.responses import JSONResponse
 
 from pycricinfo.api.endpoints.raw import router as raw_router
 from pycricinfo.api.endpoints.seasons import router as seasons_router
 from pycricinfo.api.endpoints.wrapper import router as wrapper_router
+from pycricinfo.exceptions import CricinfoAPIException
 from pycricinfo.utils import get_field_from_pyproject
 
 app = FastAPI(
@@ -20,6 +22,14 @@ app = FastAPI(
 app.include_router(wrapper_router)
 app.include_router(raw_router)
 app.include_router(seasons_router)
+
+
+@app.exception_handler(CricinfoAPIException)
+async def my_custom_exception_handler(request: Request, exc: CricinfoAPIException):
+    return JSONResponse(
+        status_code=exc.status_code,
+        content=exc.output(),
+    )
 
 
 if __name__ == "__main__":